+++ packages/vendure-plugin-stripe-subscription/src/api/graphql-schema.ts
@@ -45,6 +45,10 @@ export const shopSchemaExtensions = gql`
     stripeSubscriptionPublishableKey: String
   }
 
+  extend type OrderLine {
+    stripeSubscriptions: [StripeSubscription!]
+  }
+
   extend type Query {
     previewStripeSubscriptions(
       productVariantId: ID!