+++ packages/vendure-plugin-stripe-subscription/CHANGELOG.md
@@ -1,6 +1,10 @@
-# 2.2.0 (2023-11-06)
+# 2.2.1 (2023-11-06)
 
-- extended the `HistoryEntryList` enum to make stripe subscription  custom history entry component work
+- Extended the `HistoryEntryList` enum to make stripe subscription  custom history entry component work
+
+# 2.2.0 (2023-11-08)
+
+- Return subscriptions per order line on orders
 
 # 2.1.0 (2023-11-02)